Fix [npc] the npc don't move when player move is view on x axe

+++ include/game.h
@@ -134,7 +134,7 @@ typedef struct game_s {
     npc_t *npc;
     sfTexture *zombie_texture;
     key_struct_t key;
-    int *buffer;
+    float *buffer;
 } game_t;
 
 #endif /* GAME_H_ */
